Switch the input to uint64_t.
Since it's not possible to enter negative values switch the input type to
uint64_t so its entire range can be used for input values.

Some unit tests were adjusted to match the new behaviour. This also
removes unit tests accepting negative values as input, which the engine
doesn't support. This due to '-' is always interpreted as a binary
operator and never as an unary operator.