Allow grouping in the input.
This accepts the values _ and , as grouping values. All possible sizes of
the groups are allowed. For example: 1__,,_2 but even 1__,,_ are valid
groupings.

Note in the future there might be more restrictions, but this is a simple
way to allow inputs. The wanted grouping symbol and size may depend on the
base used. For example:
0777_777
0b000_111
0b0000_1111
0x00_00
0x0000_ffff
1,000
All these may be useful depending on the source of the input.