Code cleanups.
The new code has quite a bit of duplication. Other places in the code also
can use a bit of cleaning up.

Note during the refactoring some new bugs were discovered. Added unit
tests and fixed the bugs. This was done in the same commit since
refactoring unvealed the bugs and the lack of tests.