Switches imports to include.
When Clang uses C++ 20 modules there are issues with header units.
Switching back to includes it mkes it easier to transition to C++ 20
modules instead of Clang modules.

Note in some cases more headers are added than might be strictly needed.
This has not been optimized, since the goal is to import the std module
and have no includes at all.