Switches CI to Ninja 1.11.
This is the version of Ninja required to build modules using CMake 3.26.
This switches the build to use Ninja instead of Make. However using
parallel builds doesn't work too well, thus the build is limited to one
job. (The issue is that the CMake dependencies are hand crafted and
they seem to have some errors, these do not occur when using one job.
Using C++ modules with CMake should fix this since the tool finds the
dependencies.)